+++ example/lib/lcd.dart
@@ -9,7 +9,8 @@ import 'main.dart';
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
 
-  await SystemChrome.setPreferredOrientations([DeviceOrientation.landscapeRight, DeviceOrientation.landscapeRight]);
+  await SystemChrome.setPreferredOrientations(
+      [DeviceOrientation.landscapeRight, DeviceOrientation.landscapeRight]);
   runApp(const MyApp());
 }
 
@@ -128,7 +129,8 @@ class _HomeState extends State<Home> {
                   context,
                   labels: ['LCD Bitmap', '128x40'],
                   onPressed: () async {
-                    Uint8List byte = await readFileBytes('assets/images/128x40.png');
+                    Uint8List byte =
+                        await readFileBytes('assets/images/128x40.png');
                     await SunmiPrinter.lcdImage(byte);
                   },
                 ),
@@ -136,7 +138,8 @@ class _HomeState extends State<Home> {
                   context,
                   labels: ['LCD Bitmap', '16x16'],
                   onPressed: () async {
-                    Uint8List byte = await readFileBytes('assets/images/16x16.png');
+                    Uint8List byte =
+                        await readFileBytes('assets/images/16x16.png');
                     await SunmiPrinter.lcdImage(byte);
                   },
                 ),
@@ -144,7 +147,8 @@ class _HomeState extends State<Home> {
                   context,
                   labels: ['LCD Bitmap', 'dash.jpeg', 'But no fit.'],
                   onPressed: () async {
-                    Uint8List byte = await readFileBytes('assets/images/dash.jpeg');
+                    Uint8List byte =
+                        await readFileBytes('assets/images/dash.jpeg');
                     await SunmiPrinter.lcdImage(byte);
                   },
                 ),
@@ -152,21 +156,24 @@ class _HomeState extends State<Home> {
                   context,
                   labels: ['Fill String', '32px, fill=false'],
                   onPressed: () async {
-                    await SunmiPrinter.lcdFillString('Size32', size: 32, fill: false);
+                    await SunmiPrinter.lcdFillString('Size32',
+                        size: 32, fill: false);
                   },
                 ),
                 _buildButton(
                   context,
                   labels: ['Fill String', '16px, fill=true'],
                   onPressed: () async {
-                    await SunmiPrinter.lcdFillString('abcDEFgj0123\$&=+', size: 16, fill: true);
+                    await SunmiPrinter.lcdFillString('abcDEFgj0123\$&=+',
+                        size: 16, fill: true);
                   },
                 ),
                 _buildButton(
                   context,
                   labels: ['Fill String', '64px, fill=true'],
                   onPressed: () async {
-                    await SunmiPrinter.lcdFillString('64px', size: 64, fill: true);
+                    await SunmiPrinter.lcdFillString('64px',
+                        size: 64, fill: true);
                   },
                 ),
                 _buildButton(
@@ -205,7 +212,8 @@ class _HomeState extends State<Home> {
     );
   }
 
-  Widget _buildButton(BuildContext context, {required List<String> labels, required VoidCallback onPressed}) {
+  Widget _buildButton(BuildContext context,
+      {required List<String> labels, required VoidCallback onPressed}) {
     return ElevatedButton(
       child: Column(
         mainAxisAlignment: MainAxisAlignment.center,